+++ module-apps/application-calendar/models/DayEventsModel.hpp
@@ -4,19 +4,25 @@
 #pragma once
 #include "Application.hpp"
 #include "application-calendar/widgets/CalendarStyle.hpp"
-#include "InternalModel.hpp"
+#include "DatabaseModel.hpp"
 #include <ListItem.hpp>
 #include <ListItemProvider.hpp>
 #include <module-db/Interface/EventsRecord.hpp>
 
-class DayEventsInternalModel : public app::InternalModel<gui::ListItem *>, public gui::ListItemProvider
+class DayEventsModel : public app::DatabaseModel<EventsRecord>, public gui::ListItemProvider
 {
     app::Application *application = nullptr;
+    std::string dayMonthTitle;
+    TimePoint filterFrom = TIME_POINT_INVALID;
+    TimePoint filterTill = TIME_POINT_INVALID;
 
   public:
-    DayEventsInternalModel(app::Application *app);
+    DayEventsModel(app::Application *app);
 
-    void loadData(std::unique_ptr<std::vector<EventsRecord>> records);
+    void setFilters(TimePoint from, TimePoint till, const std::string &dayMonth);
+
+    bool updateRecords(std::vector<EventsRecord> records) override;
+    auto handleQueryResponse(db::QueryResult *) -> bool;
 
     [[nodiscard]] unsigned int getMinimalItemHeight() const override;
     [[nodiscard]] unsigned int requestRecordsCount() override;